tosa to ttir refactor: split pass and patterns into separate files
